🔥 Remove Integration Test
Fixing the integration test is not worth the effort, because the case that it tests for has been overcome by events. The integration test exists to catch errors with shading, but we no longer shade in this project, so it is unnecessary. We will similar integration tests to the cli project where they're more effective.