🚚 Integrate Language Providers as Submodules (#8)
Now includes separate modules for CLI and language providers that will are stitched together by Gradle build tool.

Packages the JavaScript language provider JS bundle into the native image and executes it as a guest language. Just a demonstration at this time. Gradle packages the JavaScript bundle into an artifact that is shared with the cli project to ultimately produce a native binary.