🎨 Automated Formatting for Gradle Plugins (#11)
The automated formatting Gradle plugin we're using is called `spotless`. It was working for the main gradle build, but it was not formatting files in the `gradle` directory as expected.

Running `./gradlew check` and `./gradlew spotlessApply` from the root will check for formatting problems and automatically-format source code respectively, for all the files including those in `gradle`.