🔥 Remove CodeTL CLI and JavaScript Langauge Provider
Since we no longer intend to build a CodeTL CLI that has both the Java
and JavaScript language providers in one GraalVM native image, this
change removes that old strategy from the codebase. This change does not
go so far as to rename all the instances of "codetl" just yet: I'm
making this smaller change now, to avoid needing to maintain the GitHub
token we use for pulling the submodules.