Added fix for an occasional bug for git repo tests (#40)
Fixes an occasional `org.eclipse.jgit.api.errors.JGitInternalException`
that was being thrown. Most likely caused by interruptions when cloning
the repository or resetting it.

You can reproduce the bug by creating a `/tmp/WebGoat820` folder with an
internal empty `.git` folder and executing the WebGoat820Test.

The solution was to just delete the folder if it's there before cloning.