Translate SSRF to codemod (#38)
Translated the SSRF visitor to a codemod. It required very few changes
to make work to inject the necessary dependency.

The existing `DependencyInjectingVisitor` honors our changes and injects
the dependencies. We will need to eventually, when we're done with the
"visitor" approach and have completely switch over to codemods, create a
new approach for doing this, but for now we're good.

---------

Co-authored-by: pixeebot[bot] <104101892+pixeebot[bot]@users.noreply.github.com>
Co-authored-by: pixee <help@pixee.ai>