Switch Apache multipart visitor to codemod (#42)
Translated the `ApacheMultipartVisitor` to a codemod.

Added `--dataflow-traces` to the Semgrep CLI invocation so that we could
get the whole dataflow in the SARIF. Some of our rules won't want the
"sink" of a flow, but instead the "source". Now we also have a way of
telling the framework which is interesting for our codemod to receive.

This also adds a critical test to our codemod testing framework. We now
run the codemod twice and ensure that the second run didn't produce any
changes. Our codemods need to be smart about this and verify it, or we
could get into a loop of constantly pushing redundant/nonsense changes.