💚 Fix Gradle Lifecycle Task Propagation (#60)
We're also feeling the pain described in
https://github.com/gradle/gradle/issues/20863

I repurposed the `io.codemodder.root` plugin to be a plugin for project
root build scripts (which makes far more sense than its previous use as
a plugin for a single build).

In the root build script for the composite build, I added lifecycle
tasks that depend on the lifecycle tasks of each included build.

Now, when users run a lifecycle task for a project root (e.g. `./gradlew
:framework:check`, the logic in `io.codemodder.root` makes sure that
works as a user would expect if this were not a composite build.
Likewise, when a user runs a lifecycle task at the root (e.g. `./gradlew
check`), then the task propagates to all included builds (as one would
expect).

I added the file versioning plugin to the root build script, so that we
can execute `./gradlew setVersion` as the release.yml does.