💚 Fix Gradle Publish Task (#62)
Running "publish" from the root should publish all artifacts in included
builds.

This fix for the "publish" task is different than the fix for the
lifecycle tasks, because "publish" is not a lifecycle task added by
Gradle's `base` plugin. Therefore, it does not exist on all projects:
only those that need to publish artifacts.

This change adds a "publish" task to any root project build script that
lacks it. The `legacy` and `codemodder-community-codemod` builds are not
mulitproject builds _and_ the root build script has a publish plugin, so
they did not need to change. The `plugin` and `framework` builds are
multi-project builds, and their root build script does not have a
`publish` task, but their subprojects need to be published. So, these
root build scripts use a new plugin `io.codemodder.publishing-root` to
signify that they are a root build script for a multi-project build that
contains subprojects that must be published.

These changes are necessary to work around
https://github.com/gradle/gradle/issues/20863. That issue describes a
problem wherein interacting with a composite build is significantly
different than interacting with a typical multi-project build.