Create a Maven plugin for the codemodder framework (#73)
This PR creates a new Maven plugin for our framework. We also (somewhat
forcefully) use this plugin from the old CLI as well as the new CLI.

This code is a lot nicer, better tested, and we can now delete a lot of
code now that we have no more `FileVisitor`s.