Add missing Semgrep fields if detected missing (#84)
The users don't use any of the metadata fields in their Semgrep rules
(`languages: java`, `severity: WARNING`, `message: Blah`). This PR adds
them to YAML if we detect they are missing so they can focus on the
patterns.

Also, if the user provides a single Semgrep rule (single ` - id: foo`),
it is gratuitous punishment for them to have to _tell us_ that it's that
rule it's interested in.

These checks and additions are simple but seem pretty damn effective at
the job of making codemods eaier to write. This does feel a smidge hacky
to me, but I don't feel they subject us to any long term technical debt.