Only write to shared build cache on main branch
This is usually how people start with the build cache,
to help minimize and isolate the potential for any issues.
Gradle usually recommends starting with only writing to the shared
build cache (in this case, a GitHub actions cache of the Gradle
build cache directory) across actions on the main branch. As time
goes on, we can and should expand build cache writes to feature
branches as well. Worth noting, it is usually not recommended to
write to shared build caches (such as a remote build cache over
http) from developer machines, as it contributes relatively little
value and adds significant opportunity for risk/problems