fix(LLMAssistedCodemod): line number in CodeTF is off by 1 (#137)
The line number of the change was being set to the position of the first
delta, but the position is 0-based. For example:
```java
Patch<String> patch = DiffUtils.diff(List.of("hello"), List.of("goodbye"));
System.out.println(patch.toString());
```

Results in a position of `0` instead of `1`:
```
Patch{deltas=[[ChangeDelta, position: 0, lines: [hello] to [goodbye]]]}
```

We now add 1 to the position to get the actual line number.