Migrate some commons-io methods to NIO (#184)
This PR is notable in that it also introduces the concept of removing an
`import` if it detects there are no references to it in the code any
longer. The next step would be to perform this same search over the
codebase and nuke the dependency, but that feels far too ambitious to
cram into this.