Move from `String`-based charset lookup to `StandardCharsets` (#183)
This codemod is straightforward but it's notable in that it attempts to
"clean up" the exception handling of the method, which may need changes
after the code change.

Compare the `Test.java.before/after` to see the types of code changes
this now supports. Happily, the type resolution provided by JavaParser
means this will generally be correct a lot. I think this is safe enough
to move to `DefaultCodemods` -- meaning it will be right 95% of the
time, and when it's not right, it's an easy fix -- but happy to hear
your opinion.