Added short-circuiting logic for codemods (#203)
Added a hook for `CodemodRunner` to short-circuit running a codemod if
the codemod _knows_ it doesn't have the pre-requisites to make any
changes. For example, if a CodeQL scan is not provided via SARIF inputs
on the command line, then there is no way a CodeQL-based codemod will
make any changes -- so we shouldn't even try.

This helps performance on all projects, but is game-breaking on
extremely big projects. For instance, on the `Grasscutter` repo, our
pre-PR testing shows it was on track for 3+ hours. After this change,
execution finishes in 25 minutes.