Batch Semgrep analysis before execution (#207)
Right now we run Semgrep for each individual Semgrep-assisted codemod.
Assuming there are 100 files and 10 Semgrep codemods, this means there
are 1,000 file reads executed in a given “scan”.

However, let’s assume that the average case is there are actually 3
“hits” for Semgrep codemods. It’s possible to design a system more
optimized to prevent extra file reads.

This “batch-and-re-run” strategy change batches all the Semgrep rules
together and runs them once before any codemod execution. So, in our
example, this means we'd run Semgrep _once_, but with all 3 rules. This
would mean 100 files would be read to find the “3 hits”. The framework
would then operate as normal, except for the 7 which didn’t hit, it
would ****not**** re-run Semgrep again.

This would mean a reduction in file reads from 1,000 to 600. I
anticipate this would bring the cost of the average case down a lot.