Remove EmbedderFacade and refactor POMScanner due to this changes (#219)
`EmbedderFacade` was an utility for invoking Maven Embedder to build and
resolve project models. It encapsulated the process of setting up the
Maven environment and resolving project models. It was used by
QueryByResolver, UnwrapEffectivePom and POMScanner.

`POMScanner` used this to retrieve the Parent POMs to set the Parent POM
Documents of the original project model and in case of any failure,`
POMScanner::legacyScanFrom` would handle the whole process. By removing
`EmbedderFacade` now it always executes ` POMScanner::legacyScanFrom`