Added resource leak transform and general codemod (#339)
Added a new resource leak transform and general codemod

- Removed CodeQL dependency for ResourceLeakFixer and left it available
as a general transform.
- Added new resource leak codemod independent of CodeQL. This may
conflict with the resource leak codemods if both are available,
producing more changes than necessary.
- Added ResourceLeakFixer transform to
PreventFileWriterLeakWithFilesCodemod. Closes #304.

You may notice the changes to the PreventFileWriterLeakWithFilesCodemod
test files. This is due to a newly discovered issue. Refer to #338.