Add fix for tab-oriented confusion  (#369)
This result for the `AddClarifyingBracesCodemod` was confusing, because
it shouldn't act on the statement. There's no visual confusion in the
single statement within the if block.

The math the codemod does for calculating the offsets involved and
deciding to change is based on column number. This makes sense if the
code is written with spaces, not tabs. With tabs, the "visual layout" of
the code depends on the viewer, and the offsets would require different
math at a minimum. Our workaround is to ignore cases that involve tabs
for now until we get a better design to accommodate tabs.

<img width="434" alt="image"
src="https://github.com/pixee/codemodder-java/assets/911610/9de959b0-d952-42a9-b673-54716a513a47">