🔇 Reduce Noise in Logs
When codemodder encounters a text file with a character encoding that it
does not support, then it fails to read that file and emits an error
level log with a stack trace. This can be very noisy.

Instead, we should emit a warn level log with no stack trace when this
occurs. warn feels right for a couple of reasons:

* this error does not prevent codemodder from completing its analysis;
  rather, codemodder adds the failed file to its "failed files" list in
  its CodeTF report.
* there will always be cases for which we have no recourse for this
  problem.

Also, we need to include the file path in the warning message.

/close #work