♻️ Better Way to Indicate Lack of Column Information
When a user of these "Remediator" or "Fixer" strategies has no column
data to provide, they can provide a null column strategy instead of
providing a strategy that maybe provides a column.

This design assumes that each tool knows whether or not there is a
column. It does not leave room for a sometimes providing column data for
a specific rule x tool combination.