Remediator API and adjacent classes refactoring (#455)
Adds a new API for the remediator/searcher pattern we use for codemods.
The aim is to eliminate a lot of duplicate/similar-ish code around the
remediators. Also converts the Sonar JNDI codemod to use the new API.

Classes from the previous API are still there, just renamed with the
`Legacy` prefix.